+++ src/main/java/be/baur/sds/model/ModelGroup.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import be.baur.sda.DataNode;
 import be.baur.sds.Component;
-import be.baur.sds.parsing.Attribute;
+import be.baur.sds.parser.Attribute;
 
 /**
  * This is the abstract superclass of all model groups.